+++ src/structs.jl
@@ -29,7 +29,7 @@ mutable struct SQLQuery
     athena_params::Any    
     limit::String
     ch_settings::String
-    
+
     function SQLQuery(;select::String="", from::String="", where::String="", groupBy::String="", orderBy::String="", having::String="", 
         window_order::String="", windowFrame::String="", is_aggregated::Bool=false, post_aggregation::Bool=false, metadata::DataFrame=DataFrame(), 
         distinct::Bool=false, db::Any=nothing, ctes::Vector{CTE}=Vector{CTE}(), cte_count::Int=0, athena_params::Any=nothing, limit::String="", 
@@ -96,7 +96,9 @@ function from_query(query::TidierDB.SQLQuery)
         db=query.db,
         ctes=[copy(cte) for cte in query.ctes],  
         cte_count=query.cte_count,
-        athena_params = query.athena_params
+        athena_params = query.athena_params,
+        limit = query.limit,
+        ch_settings = query.ch_settings
     )
     return new_query
 end