+++ src/structs.jl
@@ -28,11 +28,14 @@ mutable struct SQLQuery
     cte_count::Int
     athena_params::Any    
     limit::String
+    ch_settings::String
 
-    function SQLQuery(;select::String="", from::String="", where::String="", groupBy::String="", orderBy::String="", having::String="", window_order::String="", windowFrame::String="",
-         is_aggregated::Bool=false, post_aggregation::Bool=false, metadata::DataFrame=DataFrame(), distinct::Bool=false, db::Any=nothing, ctes::Vector{CTE}=Vector{CTE}(),
-          cte_count::Int=0, athena_params::Any=nothing, limit::String="" )
-        new(select, from, where, groupBy, orderBy, having, window_order, windowFrame, is_aggregated, post_aggregation, metadata, distinct, db, ctes, cte_count, athena_params, limit)
+    function SQLQuery(;select::String="", from::String="", where::String="", groupBy::String="", orderBy::String="", having::String="", 
+        window_order::String="", windowFrame::String="", is_aggregated::Bool=false, post_aggregation::Bool=false, metadata::DataFrame=DataFrame(), 
+        distinct::Bool=false, db::Any=nothing, ctes::Vector{CTE}=Vector{CTE}(), cte_count::Int=0, athena_params::Any=nothing, limit::String="", 
+        ch_settings::String="")
+        new(select, from, where, groupBy, orderBy, having, window_order, windowFrame, is_aggregated, 
+        post_aggregation, metadata, distinct, db, ctes, cte_count, athena_params, limit, ch_settings)
     end
 end
 
@@ -93,7 +96,9 @@ function from_query(query::TidierDB.SQLQuery)
         db=query.db,
         ctes=[copy(cte) for cte in query.ctes],  
         cte_count=query.cte_count,
-        athena_params = query.athena_params
+        athena_params = query.athena_params,
+        limit = query.limit,
+        ch_settings = query.ch_settings
     )
     return new_query
 end